Remove TestGetShareTimeout.
It should be covered by TestDownloadSharesDialTimeout and
TestDownloadSharesDownloadTimeout
